refactor(tokowaka-client): move edge-optimize control-plane under cdn/cloudfront
Addresses review nit (nit23uec): relocate the CloudFront-specific Edge Optimize
control-plane and edge runtime code into the cdn/cloudfront folder, alongside the
existing CDN clients. Pure move — public API (package root re-exports) unchanged.

  src/edge-optimize/index.js     -> src/cdn/cloudfront/index.js
  src/edge-optimize/edge-code.js -> src/cdn/cloudfront/edge-code.js
  test/edge-optimize/index.test.js -> test/cdn/cloudfront/index.test.js
